Refactor  and  to handle category validation and API request construction.
